[FIX] l10n_tw_edi_ecpay : missing neutralization
This commit adds the missing neutralization necessary for the l10n_tw_edi_ecpay
module introduced in [1]

The purpose of the standard neutralization framework is to allow us to
create database copies that will not interact with external systems in
ways that could impact the production database (or if it is not possible
to prevent the interactions, make sure that they are benign or won't
result in actual changes), or impact the customers of the operator of
the production database.

This is mainly useful to allow safe support investigation on database
duplicates.

[1] https://github.com/odoo/odoo/pull/223132

closes odoo/odoo#240975

Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>